Make copying Minisat::Vec a compile time error
Previously it was just a runtime error, and it was disabled
in an extremely dumb way. `assert(0)` stops working in Release
configurations and then the program runs straight into UB.

It should've either used `abort`, or make those methods private
(which would make it compilation error for users of vec). Given
that we suppose C++11 anyway, we can just use `= delete;` now.