Optimize vec<T> implementation by modifying internals
These changes are

* Explicitly avoid destruction for trivially destructible types

This ensures an optimization that the optimizer could also perform,
but avoids relying on it being "sufficiently smart".

* Use a trio of pointers for internal representation, instead of
a pointer and 2 sizes.

This has some performance advantages, e.g. `push` is cheaper, because
there is no longer the need to add a pointer and an int, instead the
address is already known and can be used directly.

There are also some disadvantages, such as calls to `size` being
more expensive, because `size()` now has to subtract two pointers
instead of just returning a member. This can be potentially quite
costly, because Minisat was written with manual indexing, raw loops
and `size()` based iteration.

* Added `vec<T>::empty` to avoid performance penalty from `size() == 0`
emptiness checks.

* Added iterator support to enable iterating vec<T> with a for-each loop

* Added `vec<T>::truncate`, because a common pattern in vec<T>'s usage
inside Minisat is to shrink it by N elements, where the N is computed
by a difference of indices into the vector. By moving these to iterators
we can avoid this inefficiency.

Clean-up vec::capacity